backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
groq==0.9.0
datasets==3.2.0
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.10.3